refactor: extract action_menu / reaction_picker / message_request overlays (#357)
Sixth slice of the ui.rs decomposition. Bundles three small, fully
self-contained overlays into one PR to reduce review churn:

- draw_action_menu     (~62 lines) - per-message action list
- draw_reaction_picker (~42 lines) - quick emoji reaction strip
- draw_message_request (~62 lines) - accept/delete prompt for unaccepted contacts

Each lives in its own file under ui/overlays/. QUICK_REACTIONS is now
only imported by the reaction_picker submodule, so it was dropped from
ui/mod.rs's import block.

ui/mod.rs: 4778 -> 4610 lines.
